moved useInput to useCrossword to resolve grid duplication errors
