+++ crosswordPlayer/src/hooks/useCrossword.ts
@@ -24,6 +24,8 @@ interface UseCrosswordReturn {
     onCheckClick: () => void;
     handleClueClick: (type: 'across' | 'down', number: number) => void;
     activeClue: { type: 'across' | 'down'; number: number } | null;
+    handleKeyDown: (event: KeyboardEvent) => void;
+    handleClick: (row: number, col: number) => void;
 }
 
 export interface Word {
@@ -273,12 +275,47 @@ export function useCrossword(): UseCrosswordReturn {
         return false;
     };
 
+    const handleKeyDown = (event: KeyboardEvent) => {
+        if (!activeCell) return;
+
+        const { row, col } = activeCell;
+        const key = event.key;
+
+        if (key.length === 1 && /^[a-zA-Z]$/.test(key)) {
+            // Only allow one letter at a time
+            const newGrid = [...grid];
+            newGrid[row][col].value = key.toUpperCase();
+            setGrid(newGrid);
+        } else if (key === 'ArrowUp' && row > 0) {
+            setActiveCell({ row: row - 1, col });
+        } else if (key === 'ArrowDown' && row < grid.length - 1) {
+            setActiveCell({ row: row + 1, col });
+        } else if (key === 'ArrowLeft' && col > 0) {
+            setActiveCell({ row, col: col - 1 });
+        } else if (key === 'ArrowRight' && col < grid[0].length - 1) {
+            setActiveCell({ row, col: col + 1 });
+        }
+    };
+
+    const handleClick = (row: number, col: number) => {
+        setActiveCell({ row, col });
+    };
+
+    useEffect(() => {
+        window.addEventListener('keydown', handleKeyDown);
+        return () => {
+            window.removeEventListener('keydown', handleKeyDown);
+        };
+    }, [activeCell, grid]);
+
     return {
         grid,
         isActiveCell, 
         words,
         onCheckClick,
         handleClueClick,
-        activeClue
+        activeClue,
+        handleKeyDown,
+        handleClick
     };
-};
+}
